melhorias: validacao para nao deletar cover-profile-default ao trocar de cover

+++ bitbook-api/src/modules/users/services/users.service.ts
@@ -258,7 +258,9 @@ export class UsersService {
       const uploadResult = await this.uploadsService.uploadFile(file, 'profiles/covers', 'image');
 
       if (uploadResult && user.profile.cover) {
-        await this.uploadsService.deleteFile(user.profile.cover);
+        if (!user.profile.cover.includes('default')) {
+          await this.uploadsService.deleteFile(user.profile.cover);
+        }
       }
 
       user.profile.cover = uploadResult.url;
